+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesEffect.kt
@@ -1,6 +1,7 @@
 package org.simple.clinic.teleconsultlog.prescription.medicines
 
 import org.simple.clinic.drugs.PrescribedDrug
+import org.simple.clinic.drugs.selection.custom.drugfrequency.country.DrugFrequencyChoiceItem
 import org.simple.clinic.teleconsultlog.medicinefrequency.MedicineFrequency
 import java.time.Duration
 import java.util.UUID
@@ -13,7 +14,10 @@ data class OpenEditMedicines(val patientUuid: UUID) : TeleconsultMedicinesEffect
 
 data class OpenDrugDurationSheet(val prescription: PrescribedDrug) : TeleconsultMedicinesEffect()
 
-data class OpenDrugFrequencySheet(val prescription: PrescribedDrug) : TeleconsultMedicinesEffect()
+data class OpenDrugFrequencySheet(
+    val prescription: PrescribedDrug,
+    val medicineFrequencyToFrequencyChoiceItemMap: Map<MedicineFrequency?, DrugFrequencyChoiceItem>?
+) : TeleconsultMedicinesEffect()
 
 data class UpdateDrugDuration(
     val prescribedDrugUuid: UUID,
@@ -24,3 +28,5 @@ data class UpdateDrugFrequency(
     val prescribedDrugUuid: UUID,
     val drugFrequency: MedicineFrequency
 ) : TeleconsultMedicinesEffect()
+
+object LoadDrugFrequencyChoiceItems : TeleconsultMedicinesEffect()