+++ app/src/test/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesUpdateTest.kt
@@ -9,6 +9,9 @@ import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
 import org.simple.clinic.TestData
 import org.simple.clinic.drugs.PrescribedDrug
+import org.simple.clinic.drugs.search.DrugFrequency
+import org.simple.clinic.drugs.selection.custom.drugfrequency.country.DrugFrequencyChoiceItem
+import org.simple.clinic.drugs.selection.custom.drugfrequency.country.DrugFrequencyChoiceItems
 import org.simple.clinic.teleconsultlog.medicinefrequency.MedicineFrequency
 import java.time.Duration
 import java.util.UUID
@@ -16,6 +19,14 @@ import java.util.UUID
 class TeleconsultMedicinesUpdateTest {
   private val patientUuid = UUID.fromString("134a6669-7b4c-42dd-a763-e3015b7933ff")
   private val model = TeleconsultMedicinesModel.create(patientUuid = patientUuid)
+
+  private val medicineFrequencyToFrequencyChoiceItemMap = mapOf(
+      null to DrugFrequencyChoiceItem(drugFrequency = null, label = "None"),
+      MedicineFrequency.OD to DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.OD, label = "OD"),
+      MedicineFrequency.BD to DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.BD, label = "BD"),
+      MedicineFrequency.TDS to DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.TDS, label = "TDS"),
+      MedicineFrequency.QDS to DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.QDS, label = "QDS")
+  )
 
   private val updateSpec = UpdateSpec(TeleconsultMedicinesUpdate())
 
@@ -57,17 +68,17 @@ class TeleconsultMedicinesUpdateTest {
   }
 
   @Test
-  fun `when drug frequency is clicked, then open drug frequency sheet`() {
+  fun `when drug frequency is clicked, then open drug frequency sheet with medicine frequency to frequency choice items map`() {
     val prescription = TestData.prescription(
         uuid = UUID.fromString("b9c52365-2782-4c03-95ac-1c508a11a510")
     )
 
     updateSpec
-        .given(model)
+        .given(model.medicineFrequencyToFrequencyChoiceItemMapLoaded(medicineFrequencyToFrequencyChoiceItemMap))
         .whenEvent(DrugFrequencyClicked(prescription))
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(OpenDrugFrequencySheet(prescription))
+            hasEffects(OpenDrugFrequencySheet(prescription, medicineFrequencyToFrequencyChoiceItemMap))
         ))
   }
 
@@ -126,6 +137,25 @@ class TeleconsultMedicinesUpdateTest {
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(OpenEditMedicines(patientUuid)),
+        ))
+  }
+
+  @Test
+  fun `when drug frequency choice items are loaded, then update the model with a map of medicine frequency to frequency choice items`() {
+    val drugFrequencyChoiceItems = listOf(
+        DrugFrequencyChoiceItem(drugFrequency = null, label = "None"),
+        DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.OD, label = "OD"),
+        DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.BD, label = "BD"),
+        DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.TDS, label = "TDS"),
+        DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.QDS, label = "QDS")
+    )
+
+    updateSpec
+        .given(model)
+        .whenEvent(DrugFrequencyChoiceItemsLoaded(DrugFrequencyChoiceItems(drugFrequencyChoiceItems)))
+        .then(assertThatNext(
+            hasModel(model.medicineFrequencyToFrequencyChoiceItemMapLoaded(medicineFrequencyToFrequencyChoiceItemMap)),
+            hasNoEffects()
         ))
   }
 }